Move around types for deps reason
