+++ src/dailyphoto/exif.py
@@ -4,7 +4,7 @@ from PIL import Image
 from PIL.ExifTags import Base
 from PIL.ExifTags import TAGS
 
-from .metadata import MetadataEditable
+from .types import MetadataEditable
 
 
 def print_exif(image_files: list[str]) -> int:
