+++ src/dailyphoto/metadata.py
@@ -2,42 +2,14 @@ import json
 import os
 import subprocess
 import sys
-from typing import Annotated
 
-from annotated_types import Len
-from pydantic import BaseModel
-from pydantic import ConfigDict
 from pydantic import ValidationError
 
 from . import kitty
 from .config import Config
-from .config import ShortDatetime
 from .exif import exif_to_metadata
-
-
-class MetadataEditable(BaseModel):
-    """Represents newly generated image metadata. Used for adding + editing images when fields might be unset"""
-
-    model_config = ConfigDict(extra="allow")
-    alt: str = ""
-    camera: str = ""
-    date: ShortDatetime | str | None = None
-    film: str = ""
-    subtitle: str = ""
-
-
-class Metadata(BaseModel):
-    """
-    Represents the final image metadata.
-    Used for validation and generation to ensure that all fields are filled in and have certain properties.
-    """
-
-    model_config = ConfigDict(extra="forbid")
-    alt: Annotated[str, Len(min_length=1)]
-    camera: Annotated[str, Len(min_length=1)]
-    date: ShortDatetime
-    film: Annotated[str, Len(min_length=1)]
-    subtitle: Annotated[str, Len(min_length=1)]
+from .types import Metadata
+from .types import MetadataEditable
 
 
 def get_metadata_filename(metadata_dir: str, image: str) -> str:
